Login is now working with one small bug to be resolved
