+++ src/css/base.scss
@@ -21,6 +21,7 @@ header {
   justify-content: space-between;
   align-items: center;
   padding: 8px;
+  width: 100%;
 }
 
 h2 {
@@ -55,19 +56,20 @@ span {
 button {
   height: 30px;
   width: 80px;
-  border-radius: 3px;
+  border-radius: 1px;
   border: 1px solid #333333;
   color: #333333;
   font-weight: 600;
   background-color: #E0E0E0;
   font-family: "Quicksand", sans-serif;
+  outline: none;
 
 }
 
 input {
   height: 30px;
   padding: 0;
-  border: 1px grey solid;
+  border: 1px solid #333333;
   border-radius: 1px;
   padding-left: 5px;
   outline: none;
@@ -85,7 +87,12 @@ input {
   height: 400px;
   margin-top: 80px;
   position: relative;
+  width: 55%;
   z-index: 2;
+
+
+flex-flow: row wrap;
+
 }
