+++ src/index.html
@@ -18,7 +18,7 @@
           <label>Password:</label>
           <input id="password" type="password">
         </section>
-        <button type="button" onclick="domUpdates.logIn()">Log In</button>
+        <button class="login-button" type="button">Log In</button>
       </form>
     </article>
     <main>
